!MSF:msf_find_unused_slot_chanel - now provides random free chanel. This helps to avoid cell conflicts between nbrs.
--HG--
branch : work-msf

+++ os/services/msf/msf-reserved-cell.h
@@ -91,7 +91,7 @@ void msf_reserved_release_link(tsch_link_t *cell);
 
 //==============================================================================
 long msf_find_unused_slot_offset(tsch_slotframe_t *slotframe);
-uint16_t  msf_find_unused_slot_chanel(uint16_t slot);
+int  msf_find_unused_slot_chanel(uint16_t slot);
